make the transformer eager layers the same
